+++ src/main/resources/application-local.properties
@@ -1,11 +1,9 @@
-spring.profiles.active=local
-
-spring.datasource.url=jdbc:mysql://localhost:3306/ajouevent-mysql
-spring.datasource.username=root
-spring.datasource.password=Qkr4un8v812@
+spring.datasource.url=jdbc:mysql://localhost:3306/ajouevent?useSSL=false&serverTimezone=Asia/Seoul&allowPublicKeyRetrieval=true
+spring.datasource.username=ajou
+spring.datasource.password=ajou1234
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 
-spring.jpa.hibernate.ddl-auto=create
+spring.jpa.hibernate.ddl-auto=update
 spring.jpa.show-sql=true
 
 spring.sql.init.mode=always
